feat!: Support `@cesium/engine`.
This is a feature completely independent of `cesium` support, and has no impact on regular `cesium` users. As this is a major update, it is considered a BREAKING CHANGE.
For `@cesium/engine` users, this plugin can be used by `import cesium from 'vite-plugin-cesium-build/engine'`, and the usage is identical to `import cesium from 'vite-plugin-cesium-build'`.

Refs: #6